pyproject.toml
[project]
name = "mana"
version = "0.1.0"
description = "Semantic search for command-line programs using FAISS"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "sentence-transformers>=2.2.0",
    "numpy>=1.24.0",
    "faiss-cpu>=1.7.4",
]

[project.scripts]
mana = "mana.cli:main"
